Fix request_irq() for < 2.6.19
The irq handler changes (introduced in 2.6.19, not 2.6.20) dropped
struct pt_regs from the handler prototype, they are found globally now.
This introduces the back compat for older kernels.  The handler is just
a thin layer which calls the real registered handler (all this to work
around a minor little compiler warning ;-)  Needed for device assignment
on older kernels.

Signed-off-by: Chris Wright <chrisw@redhat.com>
Signed-off-by: Avi Kivity <avi@redhat.com>