Support non-eagerFPU model for kernels < 3.7
This fills in what KVM was doing before commit 9c1c3fac53. This does not
solve the issues the commit was addressing, but it's all we can
reasonably do for older kernels.

We also need to restore how cr0.TS is handled on VMX as the eagerFPU
model is not available before 3.7.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>