add wrappers for nanosecond-based timekeeping
Timekeeping code switched from timespec- to nanosecond-based in 3.17,
and some fields of struct timekeeper thus are new or changed meaning
in that release.

When they are accessed, go through a wrapper function in
external-module-compat.c, the code is tricky enough that we do not
want it in the sync script.

The logic in kvm_get_xtime_nsec is based on the code before
upstream commit cbcf2dd3b3d4d990610259e8d878fc8dc1f17d80, but
I'm retaining most of the changes in that commit.

Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>