stubs for xsavec support
These are needed for KVM changes in 3.18.

Recent kernels added a separate feature word for XSAVE features, and KVM's
CPUID code is relying on the new definition.  Except for cpu_has_xsaves,
it's never accessing the feature itself: wrap cpu_has_xsaves with
kvm_cpu_has_xsaves, and then there is no problem with out-of-bounds
accesses.

Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>
[Jan: fix warnings if X86_FEATURE_XSAVEOPT was already correct]
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>